Added NoteColorScript and removed PianoScript
Separated Piano functinality from midi note color and replaced particle system with note color. Now whenever anote is held down, the color will be red. Whenever that key is up, the object will be black. I also the midi input from input into their functions outside of update to keep things organized. The name of the of functions is based on the octave and was placed into update to essentially do the same thing as before.